chore: move auth files to shared folder
