+++ src/app/auth/auth.module.ts
@@ -9,8 +9,8 @@ import { AuthLockedComponent } from "./auth-locked/auth-locked.component";
 import { AuthPinComponent } from "./auth-pin/auth-pin.component";
 import { AuthTouchIdComponent } from "./auth-touch-id/auth-touch-id.component";
 import { AuthComponent } from "./auth.component";
-import { AuthService } from "./auth.service";
-import { AuthState } from "./auth.state";
+import { AuthService } from "./shared/auth.service";
+import { AuthState } from "./shared/auth.state";
 
 @NgModule({
 	declarations: [
