+++ src/app/auth/shared/auth.config.ts
