+++ src/app/auth/shared/auth.types.ts
