+++ src/app/pages/delegates/delegates.ts
@@ -11,7 +11,7 @@ import { Subject } from "rxjs";
 import { switchMap, takeUntil, tap } from "rxjs/operators";
 
 import * as constants from "@/app/app.constants";
-import { AuthController } from "@/app/auth/auth.controller";
+import { AuthController } from "@/app/auth/shared/auth.controller";
 import { WalletController } from "@/app/wallet/wallet.controller";
 import { ConfirmTransactionComponent } from "@/components/confirm-transaction/confirm-transaction";
 import { Wallet, WalletKeys } from "@/models/model";
