+++ src/app/pages/wallet/wallet-dashboard/wallet-dashboard.ts
@@ -25,7 +25,7 @@ import {
 } from "rxjs/operators";
 
 import * as constants from "@/app/app.constants";
-import { AuthController } from "@/app/auth/auth.controller";
+import { AuthController } from "@/app/auth/shared/auth.controller";
 import { WalletBackupModal } from "@/app/modals/wallet-backup/wallet-backup";
 import { ConfirmTransactionComponent } from "@/components/confirm-transaction/confirm-transaction";
 import {
